client/backend/requirements.txt
fastapi
pydantic
uvicorn
dashscope #optional（使用阿里灵积时安装）
volcengine #optional(使用火山翻译时安装）
python-docx
BCEmbedding==0.1.3
langchain==0.1.0
langchain-community==0.0.9
langchain-core==0.1.7
langsmith==0.0.77
# faiss-gpu for gpu environment
faiss-cpu # for cpu-only environment
pocketbase==0.10.0
gne
selenium